Add logic to display toast
